Issues presets: init issues presets menu
