+++ lib/frontend/src/components/IssuesPresets.sass
@@ -1,0 +1,13 @@
+@import ../vars
+
+.IssuesPresets
+  background: #2c4065
+  color: #fff
+  display: flex
+  flex-flow: row wrap
+  width: 100%
+
+  ul
+    align-items: stretch
+    display: flex
+    margin: 0
